Move Rocket component into componts folder
