+++ src/pages/Rockets.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchRockets } from '../redux/rockets/rocketsSlice';
-import Rocket from './Rocket';
+import Rocket from '../components/Rocket';
 
 const Rockets = () => {
   const dispatch = useDispatch();
